Refactor screens
